Add test case for calibration coefficients handling.
