+++ pygac/tests/__init__.py
@@ -26,7 +26,8 @@
 
 from pygac.tests import (test_calibrate_pod, test_slerp, test_calibrate_klm,
                          test_pod, test_tsm, test_reader, test_io,
-                         test_angles, test_init, test_utils)
+                         test_angles, test_init, test_utils, 
+                         test_calibration_coefficients)
 import unittest
 
 
@@ -36,7 +37,7 @@ def suite():
     mysuite = unittest.TestSuite()
     tests = (test_slerp, test_calibrate_klm, test_calibrate_pod,
              test_pod, test_tsm, test_reader, test_io, test_angles,
-             test_init, test_utils)
+             test_init, test_utils, test_calibration_coefficients)
     for test in tests:
         mysuite.addTests(test.suite())
